datastore: cleanup keycompat
