+++ README.md
@@ -72,21 +72,29 @@ A few APIs were cleaned up, and there are some differences:
 * `appengine/socket` is not required on App Engine flexible environment / Managed VMs.
   Use the standard `net` package instead.
 
-## Key Encode Decode compatibiltiy to help with datastore library migrations
+### Key Encode/Decode compatibiltiy to help with datastore library migrations
 
 Key compatibility updates have been added to help customers transition from google.golang.org/appengine/datastore to cloud.google.com/go/datastore.
 The `EnableKeyConversion` enables automatic conversion from a key encoded with cloud.google.com/go/datastore to google.golang.org/appengine/datastore key type.
 
-### Enabling key conversation
+### Enabling key conversion
+
 Enable key conversion by calling `EnableKeyConversion(ctx)` in the `/_ah/startup` handler for basic and manual scaling or any handler in automatic scaling.
 
-### 1. Basic or manual scaling
+#### 1. Basic or manual scaling
+
 This startup handler will enable key conversion for all handlers in the service.
+
 ```
-    http.HandleFunc("/_ah/start", func(w http.ResponseWriter, r *http.Request) {
-		datastore.EnableKeyConversion(appengine.NewContext(r))
+  http.HandleFunc("/_ah/start", func(w http.ResponseWriter, r *http.Request) {
+    datastore.EnableKeyConversion(appengine.NewContext(r))
 	})
 ```
-### 2. Automatic scaling
-Since `/_ah/start` is not called in automatic scaling and `/_ah/warmup` is not guaranteed, a call to `datastore.EnableKeyConversion(appengine.NewContext(r))`
-is needed in each handler where key conversion is needed.  `EnableKeyConversion` is safe for concurrent use.
+
+#### 2. Automatic scaling
+
+`/_ah/start` is not supported for automatic scaling and `/_ah/warmup` is not guaranteed to run, so you must call `datastore.EnableKeyConversion(appengine.NewContext(r))`
+before you use code that needs key conversion.
+
+You may want to add this to each of your handlers, or introduce middleware where it's called.
+`EnableKeyConversion` is safe for concurrent use. Any call to it after the first is ignored.
